ccwrap.c: Uncomment extra -lm when linking libstdc++
Linking the implied libstdc++ when linking C++ programs requires libm

The was previously commented, probably because msul libc (and possibly
also uClibc) do not require specifying -lm to get the math routines.
musl supplies the math routines in the main library and provides an empty
libm.so stub which simply refers to libc.so.

For ccwrap to be a libc agnostic CC/CXX wrapper program we should
include -lm when pulling in libstdc++.