Enable build of virtio console
Added CONFIG_VIRTIO_CONSOLE to the kernel baseconfig

This allows qemu invocations such as:

     -device virtio-serial
     -chardev socket,path=/tmp/foo,server,nowait,id=foo
     -device virtconsole,chardev=foo

And gives us a nicely blocking serial port at /dev/hvc0 which
can communicate with /tmp/foo on the host. This is useful for
implementing a simple IPC allowing the host to send some commands
to be executed in the aboriginal qemu environment.