ccwrap.c: Added CCWRAP_EXTRA_LIBDIRS environment variable
When running the native compiler in aboriginal at the link stage,
the default library search paths include /usr/cc/lib and /usr/lib,
however, various packages in a full linux stack expect that /lib
be a part of the compiler's default search paths and dont provide
an -L argument for /lib.

The added environment variable allows build systems to define
extra paths in the default search paths as a flexible way to
handle this case.

This patch also shares the path searching algorithm with
the find_in_path() function, and also adds a couple of standard
defines for better readable C code (TRUE, FALSE and NULL).