ccwrap.c: Handle -pthread specially
The -pthread flag is a relatively new beast which is basically
shorthand for -lpthread at link time and -D_REENTRANT at preprocessing time.

With -pthread and -nostdlib, gcc does not insert the -lpthread line, so
packages which use the -pthread flag do not get -lpthread at link time.

This patch does not remove -pthread from the kept compiler flags, so that
the compiler will continue to declare preprocessor directives that are
expected by -pthread, but adds -lpthread at link stage.