Added CONFIG_NET_CORE to linux baseconfig
This fixes our CONFIG_VIRTIO_NET option and makes it meaningful.

This option unlocks a wealth of stuff for network drivers which
just arent available if this net core thingamajiggie is not
enabled.

Also added CONFIG_VIRTIO for good measure, even though it seems
to be implied by CONFIG_VIRTIO_PCI