Add --with-native-system-header-dir to GCC cross builds
I had problems with the "cross-built native" GCC. This is where we
build with something like '--build=x86_64 --host=armv5 --target=armv5'.
GCC's 'configure.ac' treats this like any native build, which leads to
'target_header_dir' being set to /usr/include, which is wrong.

The symptom of this that I hit was a compile failure due to missing
sys/sdt.h (part of SystemTap). GCC's configure had found this in
/usr/include, but it (correctly) wasn't available at build time.

To work around this we can use the --with-native-system-header-dir
option to tell GCC's configure script to explicitly avoid system
headers.

Note that the same ./configure invocation is used to build the
host->target cross compiler (e.g. '--build=x86_64 --host=x86_64
--target=armv5'), but it seems harmless to pass
--with-native-system-header-dir there as well.